feat: add Program component
Added a Program component which takes data from programs.json and renders the list of programs as cards in a marquee component.

Added logos for the programs Digitomize has been a part of so far.

The json file can be changed as needed in the future to add more programs.